add location entity and update section

+++ src/main/java/ru/sfedu/ictis/sports_sections/service/Impl/SectionServiceImpl.java
@@ -12,14 +12,19 @@ import ru.sfedu.ictis.sports_sections.dto.request.SectionDtoRequest;
 import ru.sfedu.ictis.sports_sections.dto.response.CategoryDtoResponse;
 import ru.sfedu.ictis.sports_sections.dto.response.GetSectionDtoResponse;
 import ru.sfedu.ictis.sports_sections.dto.response.PagenableResponse;
+import ru.sfedu.ictis.sports_sections.dto.response.UserResponse;
 import ru.sfedu.ictis.sports_sections.entity.CategoryEntity;
+import ru.sfedu.ictis.sports_sections.entity.LocationEntity;
 import ru.sfedu.ictis.sports_sections.entity.SectionEntity;
 import ru.sfedu.ictis.sports_sections.entity.UserEntity;
 import ru.sfedu.ictis.sports_sections.exception.CustomException;
 import ru.sfedu.ictis.sports_sections.exception.ErrorCodes;
 import ru.sfedu.ictis.sports_sections.mapper.CategoryMapper;
+import ru.sfedu.ictis.sports_sections.mapper.LocationMapper;
 import ru.sfedu.ictis.sports_sections.mapper.SectionMapper;
+import ru.sfedu.ictis.sports_sections.mapper.UserMapper;
 import ru.sfedu.ictis.sports_sections.repository.CategoryRepository;
+import ru.sfedu.ictis.sports_sections.repository.LocationRepository;
 import ru.sfedu.ictis.sports_sections.repository.SectionRepository;
 import ru.sfedu.ictis.sports_sections.repository.UserRepository;
 import ru.sfedu.ictis.sports_sections.service.SectionService;
@@ -39,40 +44,56 @@ public class SectionServiceImpl implements SectionService {
 
     private final CategoryRepository categoryRepository;
 
+    private final LocationRepository locationRepository;
+
     private final SectionMapper sectionMapper;
 
     private final CategoryMapper categoryMapper;
 
+    private final UserMapper userMapper;
+
+    private final LocationMapper locationMapper;
+
     public SectionServiceImpl(SectionRepository sectionRepository,
                               UserRepository userRepository,
-                              CategoryRepository categoryRepository,
+                              CategoryRepository categoryRepository, LocationRepository locationRepository,
                               SectionMapper sectionMapper,
-                              CategoryMapper categoryMapper) {
+                              CategoryMapper categoryMapper, UserMapper userMapper, LocationMapper locationMapper) {
         this.sectionRepository = sectionRepository;
         this.userRepository = userRepository;
         this.categoryRepository = categoryRepository;
+        this.locationRepository = locationRepository;
         this.sectionMapper = sectionMapper;
         this.categoryMapper = categoryMapper;
+        this.userMapper = userMapper;
+        this.locationMapper = locationMapper;
     }
-
+    // ok
     @Override
     public Long createSection(SectionDtoRequest sectionDtoRequest) {
-        UserEntity trainer = userRepository.findByEmail(getCurrentUserEmail())
+        UserEntity admin = userRepository.findByEmail(getCurrentUserEmail())
                 .orElseThrow(() -> new CustomException(ErrorCodes.USER_NOT_FOUND));
 
-        if (!trainer.getRole().equals("trainer")) {
+        if (!admin.getRole().equals("admin")) {
             throw new CustomException(ErrorCodes.NO_RIGHTS_FOR_CREATE);
         }
 
+        LocationEntity location = locationRepository.findById(sectionDtoRequest.getLocationId())
+                .orElseThrow(() -> new CustomException(ErrorCodes.LOCATION_NOT_EXISTS));
+
+        if (sectionRepository.existsByNameAndLocation(sectionDtoRequest.getName(), location)) {
+            throw new CustomException(ErrorCodes.SECTION_EXISTS);
+        }
+
         SectionEntity sectionEntity = sectionMapper.toSectionEntity(sectionDtoRequest);
-        sectionEntity.setTrainer(trainer);
+        sectionEntity.setLocation(location);
 
         Set<CategoryEntity> categoryEntities = addAndGetAllCategories(sectionDtoRequest.getCategories());
         sectionEntity.setCategoryEntities(categoryEntities);
         sectionRepository.save(sectionEntity);
         return sectionEntity.getId();
     }
-
+    // ok
     @Override
     public PagenableResponse<GetSectionDtoResponse> getSections(Integer page, Integer perPage) {
         Pageable pageable = PageRequest.of(page - 1, perPage, Sort.by(Sort.Direction.DESC, "id"));
@@ -82,25 +103,38 @@ public class SectionServiceImpl implements SectionService {
         return new PagenableResponse<>(sectionsPageToListGetSectionDto(sectionPage), sectionPage.getTotalElements());
     }
 
+    // ok
+    @Override
+    public PagenableResponse<GetSectionDtoResponse> getSectionsWithTrainer(Integer page, Integer perPage) {
+        Pageable pageable = PageRequest.of(page - 1, perPage, Sort.by(Sort.Direction.DESC, "id"));
+
+        Page<SectionEntity> sectionPage = sectionRepository.findAllWithTrainers(pageable);
+
+        return new PagenableResponse<>(sectionsPageToListGetSectionDto(sectionPage), sectionPage.getTotalElements());
+    }
+
+    // ok vrode
     @Override
     public PagenableResponse<GetSectionDtoResponse> findSections(String trainer,
-                                          String keywords,
-                                          String location,
-                                          Set<String> categories,
-                                          Integer page,
-                                          Integer perPage) {
+                                                                 String keywords,
+                                                                 String location,
+                                                                 Set<String> categories,
+                                                                 Integer page,
+                                                                 Integer perPage) {
         Pageable pageable = PageRequest.of(page - 1, perPage, Sort.by(Sort.Direction.DESC, "id"));
 
         Specification<SectionEntity> spec = Specification.where(SectionSpecification.hasTrainerName(trainer))
                 .and(SectionSpecification.hasSectionName(keywords))
                 .and(SectionSpecification.hasLocation(location))
-                .and(SectionSpecification.hasCategories(categories));
+                .and(SectionSpecification.hasCategories(categories))
+                .and(SectionSpecification.hasTrainers());
 
         Page<SectionEntity> sectionPage = sectionRepository.findAll(spec, pageable);
 
         return new PagenableResponse<>(sectionsPageToListGetSectionDto(sectionPage), sectionPage.getTotalElements());
     }
 
+    // ok
     @Override
     public GetSectionDtoResponse getSectionById(Long id) {
         SectionEntity sectionEntity = sectionRepository
@@ -108,8 +142,15 @@ public class SectionServiceImpl implements SectionService {
                 .orElseThrow(() -> new CustomException(ErrorCodes.SECTION_NOT_FOUND));
 
         GetSectionDtoResponse res = sectionMapper.toGetSectionDtoResponse(sectionEntity);
-        res.setTrainerName(sectionEntity.getTrainer().getName());
-        res.setTrainerId(sectionEntity.getTrainer().getId());
+
+        res.setLocation(locationMapper.toLocationResponseDto(sectionEntity.getLocation()));
+
+        Set<UserResponse> trainersList = sectionEntity.getTrainers()
+                .stream()
+                .map(userMapper::toUserResponse)
+                .collect(Collectors.toSet());
+
+        res.setTrainers(trainersList);
 
         Set<CategoryDtoResponse> categoriesList = sectionEntity.getCategoryEntities()
                 .stream()
@@ -120,7 +161,7 @@ public class SectionServiceImpl implements SectionService {
 
         return res;
     }
-
+    // ok
     @Override
     public PagenableResponse<GetSectionDtoResponse> getAllSectionsOfTrainer(Integer page, Integer perPage, Long id) {
         Pageable pageable = PageRequest.of(page - 1, perPage, Sort.by(Sort.Direction.DESC, "id"));
@@ -128,13 +169,21 @@ public class SectionServiceImpl implements SectionService {
         UserEntity trainer = userRepository.findById(id)
                 .orElseThrow(() -> new CustomException(ErrorCodes.USER_NOT_FOUND));
 
-        Page<SectionEntity> sectionsPage = sectionRepository.findAllByTrainer(pageable, trainer);
+        Page<SectionEntity> sectionsPage = sectionRepository.findAllByTrainersContaining(pageable, trainer);
 
         return new PagenableResponse<>(sectionsPageToListGetSectionDto(sectionsPage), sectionsPage.getTotalElements());
     }
 
+    // ok
     @Override
     public void putSection(Long id, SectionDtoRequest section) {
+        UserEntity admin = userRepository.findByEmail(getCurrentUserEmail())
+                .orElseThrow(() -> new CustomException(ErrorCodes.USER_NOT_FOUND));
+
+        if (!admin.getRole().equals("admin")) {
+            throw new CustomException(ErrorCodes.NO_RIGHT);
+        }
+
         SectionEntity sectionEntity = sectionRepository
                 .findById(id)
                 .orElseThrow(() -> new CustomException(ErrorCodes.SECTION_NOT_FOUND));
@@ -146,8 +195,16 @@ public class SectionServiceImpl implements SectionService {
         sectionRepository.save(sectionEntity);
     }
 
+    // ok
     @Override
     public void deleteSection(Long id) {
+        UserEntity admin = userRepository.findByEmail(getCurrentUserEmail())
+                .orElseThrow(() -> new CustomException(ErrorCodes.USER_NOT_FOUND));
+
+        if (!admin.getRole().equals("admin")) {
+            throw new CustomException(ErrorCodes.NO_RIGHT);
+        }
+
         sectionRepository.deleteById(id);
     }
 
@@ -180,8 +237,15 @@ public class SectionServiceImpl implements SectionService {
         return sectionsPage.getContent().stream()
                 .map(section -> {
                     GetSectionDtoResponse dto = sectionMapper.toGetSectionDtoResponse(section);
-                    dto.setTrainerName(section.getTrainer().getName());
-                    dto.setTrainerId(section.getTrainer().getId());
+
+                    dto.setLocation(locationMapper.toLocationResponseDto(section.getLocation()));
+
+                    Set<UserResponse> trainersList = section.getTrainers()
+                            .stream()
+                            .map(userMapper::toUserResponse)
+                            .collect(Collectors.toSet());
+
+                    dto.setTrainers(trainersList);
 
                     Set<CategoryDtoResponse> categoriesList = section.getCategoryEntities()
                             .stream()
